refactor(bootstrap): move GAME_OVER subscriber registration out of HostServer
publish() can no longer reject (fault-isolated via Promise.allSettled), so
both call sites fire-and-forget with void instead of relying on synchronous
publish. HostServer no longer builds BasicStatsCalculator/UnrankedMatchSaver
or touches the DI container directly; that wiring moves to index.ts, after
persistence is bootstrapped and before any socket can publish GAME_OVER.
