+++ src/edopro/room/application/FinishDuelHandler.ts
@@ -98,7 +98,7 @@ export class FinishDuelHandler {
 
 			RoomList.deleteRoom(this.room);
 
-			this.eventBus.publish(
+			void this.eventBus.publish(
 				GameOverDomainEvent.DOMAIN_EVENT,
 				new GameOverDomainEvent({
 					bestOf: this.room.bestOf,
